+++ R/default_params_doc.R
@@ -8,6 +8,8 @@
 #'   \link{get_default_tmhmm_folder}
 #' @param locatome text similar to a FASTA file, with protein names
 #'   and the locations of its amino acids
+#' @param one_protein_fasta_filename name of a FASTA file,
+#'   that contains one name and one sequence
 #' @param protein_sequence a protein sequence
 #' @param protein_sequences one or more protein sequences
 #' @param tmhmm_bin_filename path to the TMHMM binary file.
@@ -29,6 +31,7 @@ default_params_doc <- function(
   fasta_filename,
   folder_name,
   locatome,
+  one_protein_fasta_filename,
   protein_sequence,
   protein_sequences,
   tmhmm_bin_filename,