+++ R/run_tmhmm.R
@@ -1,6 +1,10 @@
-#' Run TMHMM
+#' Run TMHMM on a FASTA file with one or more proteins.
+#'
+#' The function \link{run_tmhmm_on_one_protein_fasta_file}
+#' is the function that actually calls TMHMM, on a per-protein basis.
 #' @inheritParams default_params_doc
-#' @return a character vector with the locatome
+#' @return text similar to a FASTA file, yet with the
+#'   predicted topology instead of the protein sequence
 #' @author Richèl J.C. Bilderbeek
 #' @examples
 #' if (is_tmhmm_installed()) {
@@ -11,46 +15,31 @@
 #' @export
 run_tmhmm <- function(
   fasta_filename,
+  one_protein_fasta_filename = tempfile(fileext = ".fasta"),
   folder_name = get_default_tmhmm_folder()
 ) {
   tmhmm::check_tmhmm_installation(folder_name = folder_name)
 
-  bin_path <- file.path(
-    folder_name, "tmhmm-2.0c", "bin", "decodeanhmm.Linux_x86_64"
-  )
-  options_path <- file.path(
-    folder_name, "tmhmm-2.0c", "lib", "TMHMM2.0.options"
-  )
-  model_path <- file.path(
-    folder_name, "tmhmm-2.0c", "lib", "TMHMM2.0.model"
-  )
-  cmds <- c("-f", options_path, "-modelfile", model_path)
-  text <- NA
-  suppressWarnings(
-    text <- system2(
-      command = bin_path,
-      args = cmds,
-      stdout = TRUE,
-      stderr = NULL,
-      stdin = fasta_filename
+  # All sequences
+  t <- pureseqtmr::load_fasta_file_as_tibble_cpp(fasta_filename)
+
+  # List of all texts
+  lists <- list()
+
+  n_sequences <- nrow(t)
+
+  for (i in seq_len(n_sequences)) {
+    pureseqtmr::save_tibble_as_fasta_file(
+      t = t[i, ],
+      fasta_filename = one_protein_fasta_filename
     )
-  )
-  if (length(text) == 0) {
-    if (is.null(attr(text, "status"))) {
-      stop("Protein sequence must have at least one character")
-    }
-    testthat::expect_equal(attr(text, "status"), 100)
-    suppressWarnings(
-      text <- system2(
-        command = bin_path,
-        args = cmds,
-        stdout = NULL,
-        stderr = TRUE,
-        stdin = fasta_filename
-      )
+    text <- run_tmhmm_on_one_protein_fasta_file(
+      one_protein_fasta_filename = one_protein_fasta_filename,
+      folder_name = folder_name
     )
-    stop(text[6])
+    lists[[i]] <- tibble::tibble(text = text)
   }
-  text <- text[text != ""]
-  stringr::str_remove(string = text, pattern = "^\\?0 ")
+
+  all_text <- dplyr::bind_rows(lists)
+  all_text$text
 }