+++ tests/testthat/test-run_tmhmm_to_file.R
@@ -17,3 +17,47 @@ test_that("use", {
   expect_equal(sum(nchar(fasta_text)), sum(nchar(tmhmm_text)))
 
 })
+
+test_that("Not all sequences get their topology predicted for bigger files?", {
+  skip("Issue 4. Issue #4")
+  if (!is_tmhmm_installed()) return()
+  expect_true(is_tmhmm_installed())
+  fasta_gz_filename <- tempfile(fileext = "_UP000005640_9606.fasta.gz")
+  download.file(
+    url = "ftp://ftp.ebi.ac.uk/pub/databases/reference_proteomes/QfO/Eukaryota/UP000005640_9606.fasta.gz",
+    destfile = fasta_gz_filename
+  )
+  fasta_filename <- tempfile(fileext = "_UP000005640_9606.fasta")
+  R.utils::gunzip(
+    filename = fasta_gz_filename,
+    destname = fasta_filename,
+    remove = FALSE
+  )
+  expect_true(file.exists(fasta_filename))
+  # We know this reference proteome has 20600 proteins
+  expect_equal(
+    20600,
+    nrow(pureseqtmr::load_fasta_file_as_tibble_cpp(fasta_filename))
+  )
+  tmhmm_filename <- tempfile(fileext = "_UP000005640_9606.tmhmm")
+
+  # The original file does not work
+  expect_error(
+    run_tmhmm_to_file(
+      fasta_filename = fasta_filename,
+      tmhmm_filename = tmhmm_filename
+    ),
+    "Character 'U' not allowed in alphabet 'ACDEFGHIKLMNPQRSTVWYBXZ'."
+  )
+
+
+  skip("HIERO")
+  # Remove the sequences with U
+  t <- pureseqtmr::load_fasta_file_as_tibble_cpp()
+
+  expect_true(file.exists(tmhmm_filename))
+  expect_equal(
+    20600,
+    nrow(pureseqtmr::load_fasta_file_as_tibble_cpp(tmhmm_filename))
+  )
+})